Slim integration tests down to moto only
